Implimented MCWF in C++ for speed
MCWF was too intensive to be run easily in python, with even a 1000 iterations per measurement frequency resulting in long computational times. Instead, we re-wrote this code in C++, resulting in a significantly improved run time, allowing us to test more measurement frequencies for 20 times the time steps, while still completing the calculation in a fraction of the time.